notebook-prototype/requirements.txt
climb-onyx-client
ipywidgets==7.7.1
matplotlib
numpy
pyzmq
